bugfix - validate borrowing contracts in the Oven suite (#1455)
