+++ src/backend/ir/codegen.rs
@@ -8067,4 +8067,44 @@ def read() -> None:
             "expected checked composed-newtype bridge in generated module:\n{code}"
         );
     }
+    /// Exercise real TOML receiver facts under the library test root's declared registry-source authority.
+    #[cfg(feature = "rust_inspect")]
+    #[test]
+    fn toml_traversal_uses_extracted_receiver_contracts() -> Result<(), Box<dyn std::error::Error>> {
+        let source = r#"
+from rust::toml_edit import Item
+
+def traverse(item: Item, keys: list[str]) -> Item:
+    mut current = item
+    for key in keys:
+        match current.get(key):
+            Some(child) => current = child
+            None => return Item.None
+    return current.clone()
+
+pub def observe(item: Item) -> Item:
+    return traverse(item, ["project", "count"])
+"#;
+        let tokens =
+            crate::frontend::lexer::lex(source).map_err(|errors| std::io::Error::other(format!("lex: {errors:?}")))?;
+        let ast = crate::frontend::parser::parse(&tokens)
+            .map_err(|errors| std::io::Error::other(format!("parse: {errors:?}")))?;
+        let mut checker = crate::frontend::typechecker::TypeChecker::new();
+        let manifest_dir = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        checker.set_rust_inspect_manifest_dir(manifest_dir);
+        checker
+            .check_program(&ast)
+            .map_err(|errors| std::io::Error::other(format!("check: {errors:?}")))?;
+        let ir = crate::backend::ir::AstLowering::new_with_type_info(checker.type_info().clone())
+            .lower_program(&ast)
+            .map_err(|error| std::io::Error::other(format!("lower: {error:?}")))?;
+        let generated = crate::backend::ir::IrEmitter::new(&ir.function_registry).emit_program(&ir)?;
+        assert!(
+            generated.contains("item: &Item"),
+            "contracts: {:?}\n{generated}",
+            checker.type_info().rust.receiver_contracts
+        );
+        assert!(!generated.contains("child.clone()"), "{generated}");
+        Ok(())
+    }
 }
